fix it all
